+++ components/UIPopupStruct.m
@@ -1,6 +1,14 @@
 classdef UIPopupStruct < HandlePlus
-    
-    
+%UIPOPUPSTRUCT - Similar to UIPopup except that each item in the
+%pulldown represents a structure rather than a char.  The idea is
+%that calling val() returns a structure with extra information about
+%the system state that the pulldown represetns.  It is a convenient
+%way to lump extra information into the pulldown.  For instance you
+%might want the pulldown to represent a set of saved positions of a
+%motor but you want labels to say "Filter 1", "Filter 2" and have
+%values stored internally as "234.21" "255.22".  If you want to do
+%thatm use UIPopupStruct instead of UIPopup
+       %
     
     properties (Constant)
         dHeight = 30;
@@ -8,25 +16,41 @@ classdef UIPopupStruct < HandlePlus
     
     properties
         
-        u8Selected      % get selected index (force uint8)
-        % cSelected
+        % {uint8 1x1} the active / selected index
+        u8Selected      
 
     end
     
     properties (SetAccess = private)
         
-        cName
     end
     
     properties (Access = private)
         
+        % {cell 1xm of struct} list of structures the pulldown represents.
+        % Each struct can nave any number of fields.  Default defined in
+        % constructor
+        ceOptions
+        
+        % {char 1xm} the field of the option structure to use for the text
+        % in the pulldown
+        cField = 'cLabel';     
+        
+        % {char 1xm} the label of the pulldown
+        cLabel = '';
+        
+        % {logical 1x1} show the label?        
+        lShowLabel = false;
+        
+        % {char 1xm} the tooltip
+        cTooltip = 'Tooltip: set me!';
+        
+        % {logical 1x1} show the tooltip?
+        lShowTooltip = false;
+        
+       
         hLabel
         hUI
-        cLabel
-        lShowLabel
-        ceOptions       % 1xm cell array of struct
-        cField
-        cTooltip = 'Tooltip: set me!';
     end
     
     
@@ -36,62 +60,37 @@ classdef UIPopupStruct < HandlePlus
     
     
     methods
-        
+                       
+       function this= UIPopupStruct(varargin)
        
-               
-       function this= UIPopupStruct(stParams)
-       %UIPOPUPSTRUCT - Similar to UIPopup except that each item in the
-       %pulldown represents a structure rather than a char.  The idea is
-       %that calling val() returns a structure with extra information about
-       %the system state that the pulldown represetns.  It is a convenient
-       %way to lump extra information into the pulldown.  For instance you
-       %might want the pulldown to represent a set of saved positions of a
-       %motor but you want labels to say "Filter 1", "Filter 2" and have
-       %values stored internally as "234.21" "255.22".  If you want to do
-       %thatm use UIPopupStruct instead of UIPopup
-       %
-       % Important to initialize the structure without using
-       % struct(field1,val1, ...) constructor due to a weird way that it handles
-       % values that are cell arrays.
-       % You want to do something like:
-       %
-       % stOptionA = struct();
-       % stOptionA.name = "Test A";
-       % stOptionA.value = 100;
-       %
-       % stOptionB = struct();
-       % stOptionB.name = "Test B";
-       % stOptionB.value = 110;
-       %
-       % ceOptions = {stOptionA, stOptionB}
-       %    @param {struct} stParams - the config paramaters
-       %    @param {cell 1xm of struct} stParams.ceOptions - list of structures the pulldown
-       %        represents.   Each struct can nave any number of fields
-       %    @param {char 1xm} [stParams.cField = 'cLabel'] - the field of the option structure to use
-       %        for the text in the pulldown
-       %    @param {char 1xm} [stParams.cLabel = 'Change Me'] - the label of the pulldown
-       %    @param {logical 1x1} [stParams.lShowLabel = true] - show the label if true
-       %    @param {char 1xm} [stParams.cName = 'Default'] - the name for use in msg and logging  
-                  
-            stDefault = struct();
-            stDefault.ceOptions{1}.cLabel = 'Item 1';
-            stDefault.ceOptions{1}.dVal = 1;
-            stDefault.ceOptions{2}.cLabel = 'Item 2';
-            stDefault.ceOptions{2}.dVal = 2;
-            stDefault.ceOptions{3}.cLabel = 'Item 3';
-            stDefault.ceOptions{3}.dVal = 3;
-            stDefault.cField = 'cLabel';
-            stDefault.cLabel = 'Change Me';
-            stDefault.cName = 'Default';
-            stDefault.lShowLabel = true;
+            % Default ceOptions
+       
+            stOption1 = struct();
+            stOption1.cLabel = 'Item 1';
+            stOption1.dVal = 1;
             
+            stOption2 = struct();
+            stOption2.cLabel = 'Item 2';
+            stOption2.dVal = 2;
             
-            stParams = mergestruct(stDefault, stParams);
+            %{
+            CANNOT assign as shown below due to a quirk in the way that
+            the struct() constructor handles values that cell arrays 
+            stParams = struct(...
+                'ceOptions', ceOptions, ...
+                'cLabel', 'This Popup' ...
+            );
+            %}
             
-            this.ceOptions = stParams.ceOptions;
-            this.cField = stParams.cField;
-            this.cLabel = stParams.cLabel;
-            this.lShowLabel = stParams.lShowLabel;
+            this.ceOptions = {stOption1, stOption2};
+            
+            for k = 1 : 2: length(varargin)
+                % this.msg(sprintf('passed in %s', varargin{k}));
+                if isprop(this, varargin{k})
+                    this.msg(sprintf('settting %s', varargin{k}), 6);
+                    this.(varargin{k}) = varargin{k + 1};
+                end
+            end
             
                         
        end