+++ tests/test_keithley6482.m
@@ -1,13 +1,10 @@
-purge
 
 [cPath, cName, cExt] = fileparts(mfilename('fullpath'));
 
-% Add core
-addpath(genpath(fullfile(cPath, [sprintf('..%s', filesep), 'ui-core'])));
+% Add mic
+addpath(genpath(fullfile(cPath, '..', '..', 'mic')));
 
-% Add functions
-addpath(genpath(fullfile(cPath, [sprintf('..%s', filesep), 'functions'])));
-
+purge
 
 h = figure( ...
     'Position', [200 200 800 200] ...